rpc: move netbuf to rpc/types.h for concrete type
In order to track metrics per method it is convenient for the method to
be a structure which will contain a handler and metrics information.
However, currently the method is a type alias and in the generated code
it is made concrete. Unfortunately, in order to convert the method into
a structure we cannot get away with forward declaration of netbuf. So we
move it into rpc/types.h. There weren't any existing users of only the
netbuf.
