+++ src/v/rpc/test/netbuf_tests.cc
@@ -7,7 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "rpc/netbuf.h"
 #include "rpc/parse_utils.h"
 
 #include <seastar/core/thread.hh>
