+++ src/v/rpc/types.h
@@ -11,6 +11,7 @@
 
 #pragma once
 
+#include "bytes/iobuf.h"
 #include "likely.h"
 #include "outcome.h"
 #include "seastarx.h"
@@ -36,7 +37,6 @@
 #include <vector>
 
 namespace rpc {
-class netbuf;
 
 using clock_type = ss::lowres_clock;
 using duration_type = typename clock_type::duration;
@@ -167,6 +167,42 @@ public:
 private:
     std::vector<ss::semaphore_units<>> _reservations;
 };
+
+class netbuf {
+public:
+    /// \brief used to send the bytes down the wire
+    /// we re-compute the header-checksum on every call
+    ss::scattered_message<char> as_scattered() &&;
+
+    void set_status(rpc::status);
+    void set_correlation_id(uint32_t);
+    void set_compression(rpc::compression_type c);
+    void set_service_method_id(uint32_t);
+    void set_min_compression_bytes(size_t);
+    iobuf& buffer();
+
+private:
+    size_t _min_compression_bytes{1024};
+    header _hdr;
+    iobuf _out;
+};
+
+inline iobuf& netbuf::buffer() { return _out; }
+inline void netbuf::set_compression(rpc::compression_type c) {
+    vassert(
+      c >= compression_type::min && c <= compression_type::max,
+      "invalid compression type: {}",
+      int(c));
+    _hdr.compression = c;
+}
+inline void netbuf::set_status(rpc::status st) {
+    _hdr.meta = std::underlying_type_t<rpc::status>(st);
+}
+inline void netbuf::set_correlation_id(uint32_t x) { _hdr.correlation_id = x; }
+inline void netbuf::set_service_method_id(uint32_t x) { _hdr.meta = x; }
+inline void netbuf::set_min_compression_bytes(size_t min) {
+    _min_compression_bytes = min;
+}
 
 /// \brief most method implementations will be codegenerated
 /// by $root/tools/rpcgen.py
